exercicio 1011

+++ E1011/Main.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        int raio = scanner.nextInt();
+        double raio = scanner.nextDouble();
         final double pi = 3.14159;
         double volume = (4 / 3.0) * pi * Math.pow(raio, 3);
